move backdrop actions and reducers to redux folder
